refactor(braces): extract shared BraceUtils.wrapInBlock utility
Addresses review feedback about code duplication in brace recipes.
